docs: add documentation for custom_add
This used to have intentional typos, but they were removed.
Also subsequent commits were merged into this change,
in the sense that you can incrementally commit locally
while iterating on the change you are working on.

The fixup commit did have merge conflicts when re-ordered;
however, that is kept to showcase that sometimes there are
trade-offs that can occur when you deal with re-ordering
history if it has to be done. There's a chance this could
have been avoided if there were an additional line of space
in the comment string so that the git diff would not clobber
the two changes together.